test to get data for total number of processes over multiple nodes
this script uses write2 to first make sure all of the possible dat
files are created. It will then copy the correct file into HPL.dat and
run xhpl with the correct number of processes. If multiple nodes are
being used, then the hostile will also have to be created where the
earlier nodes are favored to have the most processes.